requirements.txt
streamlit
spacy
pdfplumber
python-docx
scikit-learn
pandas
numpy